Fixed some minor problems
In the Store I made it so that you can't sell if you don't have anything to sell, previously the logic was wrong (comparing to null when it should have been comparing to 0).
In terms of Blackjack I made it so that the player's balance information is saved after a tie or a win.
I also added an else statement after 
if (dealer.getBestTotal() == -1) on line 140 of BlackjackGame, because it was skipping the statement and displaying | Dealer: -1 You Win!, rather than what was expected |Dealer: Busted You Win!